Rework configs for the client API and the test files to remove uneeded values and move to new config types
