+++ csv/koinly_test.go
@@ -12,7 +12,7 @@ import (
 )
 
 func TestKoinlyOsmoLPParsing(t *testing.T) {
-	cfg := config.Config{}
+	cfg := config.IndexConfig{}
 	cfg.Lens.ChainID = osmosis.ChainID
 	parser := GetParser(koinly.ParserKey)
 	parser.InitializeParsingGroups()
@@ -51,7 +51,7 @@ func TestKoinlyOsmoLPParsing(t *testing.T) {
 }
 
 func TestKoinlyOsmoRewardParsing(t *testing.T) {
-	cfg := config.Config{}
+	cfg := config.IndexConfig{}
 	cfg.Lens.ChainID = osmosis.ChainID
 	parser := GetParser(koinly.ParserKey)
 	parser.InitializeParsingGroups()
