+++ test/indexer_test.go
@@ -4,7 +4,6 @@ import (
 	"os"
 	"testing"
 
-	configUtils "github.com/DefiantLabs/cosmos-indexer/config"
 	"github.com/DefiantLabs/cosmos-indexer/csv"
 	"github.com/DefiantLabs/cosmos-indexer/db"
 )
@@ -21,16 +20,11 @@ INNER JOIN addresses as addr ON addr.id = tx.sender_address_id OR addr.id = tx.r
 where addr.address = 'osmo...'
 */
 func TestOsmosisCsvForAddress(t *testing.T) {
-	config, err := configUtils.GetConfig("../config.toml")
-	if err != nil {
-		t.Fatalf("Error getting config. Err: %v", err)
-	}
-
 	addressRegex := OsmosisAddressRegex
 	addressPrefix := OsmosisAddressPrefix
 	gorm, _ := dbSetup(addressRegex, addressPrefix)
 	address := "osmo14mmus5h7m6vkp0pteks8wawaj4wf3sx7fy3s2r" // local test key address
-	csvRows, headers, err := csv.ParseForAddress([]string{address}, nil, nil, gorm, "accointing", config)
+	csvRows, headers, err := csv.ParseForAddress([]string{address}, nil, nil, gorm, "accointing")
 	if err != nil || len(csvRows) == 0 {
 		t.Fatal("Failed to lookup taxable events")
 	}
@@ -51,18 +45,13 @@ func TestOsmosisCsvForAddress(t *testing.T) {
 }
 
 func TestCsvForAddress(t *testing.T) {
-	config, err := configUtils.GetConfig("../config.toml")
-	if err != nil {
-		t.Fatalf("Error getting config. Err: %v", err)
-	}
-
 	addressRegex := "juno(valoper)?1[a-z0-9]{38}"
 	addressPrefix := "juno"
 	gorm, _ := dbSetup(addressRegex, addressPrefix)
 	// address := "juno1mt72y3jny20456k247tc5gf2dnat76l4ynvqwl"
 	// address := "juno130mdu9a0etmeuw52qfxk73pn0ga6gawk4k539x" // strangelove's delegator
 	address := "juno1m2hg5t7n8f6kzh8kmh98phenk8a4xp5wyuz34y" // local test key address
-	csvRows, headers, err := csv.ParseForAddress([]string{address}, nil, nil, gorm, "accointing", config)
+	csvRows, headers, err := csv.ParseForAddress([]string{address}, nil, nil, gorm, "accointing")
 	if err != nil || len(csvRows) == 0 {
 		t.Fatal("Failed to lookup taxable events")
 	}
